* Merge LoadRequest and GetRequest.

* Rename to ImageRequest.

* Lint warnings.

* Fix formatting.

* Fix request delegate.

* Add transformations to event listener.

* Fix method names.

* Ensure observer is cleaned up.

* Add note.

* Fix rebase.

* Fix lint.

* Update API.

* Fix tests.

* Fix typealiases.

* WIP PoolableTargetRequestDelegate.

* Change name.

* Fix import.

* Update constructor.

* Refactor.

* Refactor.

* Docs.

* Fix compilation.

* Add Request typealias.

* Remove PoolableTarget changes.

* Update API.

* Only await lifecycle if we're enqueued.

* Remove suppress.

* Remove extra return.

* Docs.

* Docs.

* Docs.

+++ coil-base/src/test/java/coil/memory/MemoryCacheTest.kt
@@ -19,7 +19,7 @@ class MemoryCacheTest {
         val weakMemoryCache = EmptyWeakMemoryCache
         val pool = BitmapPool(Int.MAX_VALUE)
         val counter = BitmapReferenceCounter(weakMemoryCache, pool, null)
-        val cache = MemoryCache(weakMemoryCache, counter, (2 * DEFAULT_BITMAP_SIZE).toInt(), null)
+        val cache = MemoryCache(weakMemoryCache, counter, (2 * DEFAULT_BITMAP_SIZE), null)
 
         val bitmap = createBitmap()
         cache.set(Key("1"), bitmap, false)
@@ -32,7 +32,7 @@ class MemoryCacheTest {
         val weakMemoryCache = EmptyWeakMemoryCache
         val pool = BitmapPool(Int.MAX_VALUE)
         val counter = BitmapReferenceCounter(weakMemoryCache, pool, null)
-        val cache = MemoryCache(weakMemoryCache, counter, (2 * DEFAULT_BITMAP_SIZE).toInt(), null)
+        val cache = MemoryCache(weakMemoryCache, counter, (2 * DEFAULT_BITMAP_SIZE), null)
 
         val first = createBitmap()
         cache.set(Key("1"), first, false)
@@ -64,7 +64,7 @@ class MemoryCacheTest {
         val weakMemoryCache = RealWeakMemoryCache()
         val pool = BitmapPool(Int.MAX_VALUE)
         val counter = BitmapReferenceCounter(weakMemoryCache, pool, null)
-        val cache = MemoryCache(weakMemoryCache, counter, (2 * DEFAULT_BITMAP_SIZE).toInt(), null)
+        val cache = MemoryCache(weakMemoryCache, counter, (2 * DEFAULT_BITMAP_SIZE), null)
 
         val bitmap = createBitmap()
         cache.set(Key("1"), bitmap, false)
@@ -78,7 +78,7 @@ class MemoryCacheTest {
         val weakMemoryCache = RealWeakMemoryCache()
         val pool = BitmapPool(Int.MAX_VALUE)
         val counter = BitmapReferenceCounter(weakMemoryCache, pool, null)
-        val cache = MemoryCache(weakMemoryCache, counter, DEFAULT_BITMAP_SIZE.toInt(), null)
+        val cache = MemoryCache(weakMemoryCache, counter, DEFAULT_BITMAP_SIZE, null)
 
         val first = createBitmap()
         cache.set(Key("1"), first, false)
@@ -98,7 +98,7 @@ class MemoryCacheTest {
         val weakMemoryCache = RealWeakMemoryCache()
         val pool = BitmapPool(Int.MAX_VALUE)
         val counter = BitmapReferenceCounter(weakMemoryCache, pool, null)
-        val cache = MemoryCache(weakMemoryCache, counter, DEFAULT_BITMAP_SIZE.toInt(), null)
+        val cache = MemoryCache(weakMemoryCache, counter, DEFAULT_BITMAP_SIZE, null)
 
         val first = createBitmap()
         cache.set(Key("key"), first, false)
